Document docutils dependency.
The configuration used in the sample rquires the rst2html filter to
be available, which in turn requires the docutils package to be
installed. Steps for installation have been added to the README.

Furthermore, a non-frozen requirements.txt file has been added for
easy Python based dependency installation. Requirements should be
frozen at some point, but given that this is a template, you can
probably argue against it.